Remove 'inline' from two methods.
The methods are only defined in the cpp and thus can't always be inlined,
the methods are virtual and so can only be inlined when the concrete type
is known, and inline functions need their definition available in all
translation units.

So just remove the 'inline'.

BUG=none
R=bo_xu@foxitsoftware.com

Review URL: https://codereview.chromium.org/409253004