Revert 'gn: cross-platform exception disable'
This reverts commit 4923e3cfbc2b617614858c427fa87a8c67aca784.

Since exceptions are in the process of being removed,
and the code currently isn't rollable into pdfium (for other
reasons) I'm going to revert this for now, so that this CL
doesn't become blocking-for-rolls if the other min/max problem
is addressed.

And, hopefully by the time I get back to this it won't be
necessary anyway.

BUG=pdfium:28,pdfium:31,chromium:354261
R=thakis@chromium.org

Review URL: https://codereview.chromium.org/432243002