Add a null check before getting unicode text in CPDF_FormField::GetValue
The test pdf file defines an invalid dictionary object with a NULL arrary
in the filed of "/V". It causes that a NULL object is returned when trying
to get the first element of this arrary. So it needs to check whether the
returned object is NULL.

BUG=395986
R=tsepez@chromium.org

Review URL: https://codereview.chromium.org/478183002